+++ Automate/Framework/LocationFloodFillIndex.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using Microsoft.Xna.Framework;
+using StardewModdingAPI;
 using StardewValley;
 using StardewValley.Buildings;
 using StardewValley.Locations;
@@ -17,8 +19,14 @@ namespace Pathoschild.Stardew.Automate.Framework
         /*********
         ** Accessors
         *********/
+        /// <summary>The log keys for entities which throw an exception when scanned, to avoid logging them repeatedly.</summary>
+        private static readonly HashSet<string> LoggedErrorKeys = new();
+
         /// <summary>The indexed entities.</summary>
         private readonly IDictionary<Vector2, object[]> Entities;
+
+        /// <summary>The monitor with which to log errors.</summary>
+        private readonly IMonitor Monitor;
 
 
         /*********
@@ -26,8 +34,11 @@ namespace Pathoschild.Stardew.Automate.Framework
         *********/
         /// <summary>Construct an instance.</summary>
         /// <param name="location">The location to index.</param>
-        public LocationFloodFillIndex(GameLocation location)
+        /// <param name="monitor">The monitor with which to log errors.</param>
+        public LocationFloodFillIndex(GameLocation location, IMonitor monitor)
         {
+            this.Monitor = monitor;
+
             this.Entities = this.Scan(location)
                 .GroupBy(group => group.Key)
                 .ToDictionary(
@@ -61,21 +72,30 @@ namespace Pathoschild.Stardew.Automate.Framework
 
             // furniture
             foreach (Furniture furniture in location.furniture)
-                yield return new(furniture.TileLocation, furniture);
+            {
+                if (!this.TryGetData("furniture", location, null, () => furniture.TileLocation, out Vector2? tile))
+                    continue;
+
+                yield return new(tile.Value, furniture);
+            }
 
             // terrain features
             foreach ((Vector2 originTile, TerrainFeature feature) in location.terrainFeatures.Pairs)
             {
-                Rectangle box = this.AbsoluteToTileArea(feature.getBoundingBox(originTile));
-                foreach (Vector2 tile in this.GetTilesIn(box))
+                if (!this.TryGetData("terrain feature", location, originTile, () => this.GetTilesIn(this.AbsoluteToTileArea(feature.getBoundingBox(originTile))), out Vector2[]? tiles))
+                    continue;
+
+                foreach (Vector2 tile in tiles)
                     yield return new(tile, feature);
             }
 
             // large terrain features
             foreach (LargeTerrainFeature feature in location.largeTerrainFeatures)
             {
-                Rectangle box = this.AbsoluteToTileArea(feature.getBoundingBox());
-                foreach (Vector2 tile in this.GetTilesIn(box))
+                if (!this.TryGetData("large terrain feature", location, null, () => this.GetTilesIn(this.AbsoluteToTileArea(feature.getBoundingBox())), out Vector2[]? tiles))
+                    continue;
+
+                foreach (Vector2 tile in tiles)
                     yield return new(tile, feature);
             }
 
@@ -122,6 +142,34 @@ namespace Pathoschild.Stardew.Automate.Framework
                 area.Width / Game1.tileSize,
                 area.Height / Game1.tileSize
             );
+        }
+
+        /// <summary>Get information from an entity, or log an error if it throws an exception.</summary>
+        /// <typeparam name="TData">The type of the data to retrieve.</typeparam>
+        /// <param name="label">A human-readable label for the entity type (like 'furniture' or 'terrain feature').</param>
+        /// <param name="location">The location containing the entity.</param>
+        /// <param name="tile">The entity's origin tile, if known.</param>
+        /// <param name="get">Read the entity data.</param>
+        /// <param name="data">The read data, if successfully read.</param>
+        /// <returns>Returns whether the data was successfully read.</returns>
+        private bool TryGetData<TData>(string label, GameLocation location, Vector2? tile, Func<TData> get, [NotNullWhen(true)] out TData? data)
+        {
+            try
+            {
+                data = get();
+                return data != null;
+            }
+            catch (Exception ex)
+            {
+                if (LocationFloodFillIndex.LoggedErrorKeys.Add($"{label}|{location.NameOrUniqueName}|{tile}"))
+                {
+                    this.Monitor.Log($"Failed accessing broken {label} in {location.NameOrUniqueName}{(tile.HasValue ? $" at {tile.Value}" : "")}. See the SMAPI log for details.", LogLevel.Warn);
+                    this.Monitor.Log(ex.ToString());
+                }
+
+                data = default;
+                return false;
+            }
         }
     }
 }